requirements.txt
torch
torchvision
numpy
pandas
pillow
matplotlib
seaborn
scikit-learn
gradio
tqdm